+++ dl/utils.py
@@ -12,7 +12,7 @@ from catalyst.contrib.criterion import CRITERION
 from catalyst.contrib.optimizers import OPTIMIZERS
 from catalyst.dl.callbacks import CALLBACKS
 from catalyst.data.dataset import ListDataset
-from catalyst.utils.fp16 import Fp16Wrap
+from catalyst.dl.fp16 import Fp16Wrap
 
 
 class UtilsFactory:
@@ -179,36 +179,6 @@ class UtilsFactory:
         for key, value in stage_config.get("args", {}).items():
             setattr(args, key, value)
         return args
-
-    @staticmethod
-    def process_epoch_metrics(
-        epoch_metrics,
-        best_metrics,
-        valid_loader="valid",
-        main_metric="loss",
-        minimize=True
-    ):
-        valid_metrics = epoch_metrics[valid_loader]
-        is_best = True \
-            if best_metrics is None \
-            else (minimize != (
-                valid_metrics[main_metric] > best_metrics[main_metric]))
-        best_metrics = valid_metrics if is_best else best_metrics
-        return best_metrics, valid_metrics, is_best
-
-    @staticmethod
-    def get_val_from_metric(metric_value):
-        if isinstance(metric_value, (int, float)):
-            pass
-        elif torch.is_tensor(metric_value):
-            metric_value = metric_value.item()
-        else:
-            metric_value = metric_value.value()
-            if isinstance(metric_value, (tuple, list)):
-                metric_value = metric_value[0]
-            if torch.is_tensor(metric_value):
-                metric_value = metric_value.item()
-        return metric_value
 
     @staticmethod
     def pack_checkpoint(
