+++ rl/algorithms/sac.py
@@ -1,5 +1,5 @@
 import torch
-from catalyst.utils.factory import UtilsFactory
+from catalyst.dl.utils import UtilsFactory
 import catalyst.rl.networks.agents as agents
 from catalyst.rl.algorithms.base import BaseAlgorithm, soft_update
