+++ rl/algorithms/td3.py
@@ -1,6 +1,6 @@
 import copy
 import torch
-from catalyst.utils.factory import UtilsFactory
+from catalyst.dl.utils import UtilsFactory
 import catalyst.rl.networks.agents as agents
 from catalyst.rl.algorithms.base import BaseAlgorithm, soft_update
